Adding support for outputing a mesh with versions
Note that inputing a mesh with versions should work fine.
Output of ipmap is not supported.
A standard mesh can be forced to be output with versions by setting the FieldVersionFlagSet. However the correct way to create a mesh with versions is to manually set the node derivative version to be used for each local mesh element node using MeshElementsNodeVersionSet. Once MeshesCreateFinish is called it will automatically calculate the total number of versions at each node in a similar manner to OpenCMISS.